More warning cleanup, this time in kernel/arch/dreamcast.
Same deal as the last commit: most of these are signed vs. unsigned comparisons.

+++ kernel/arch/dreamcast/hardware/maple/maple_utils.c
@@ -131,7 +131,7 @@ static const char *maple_resp_names[] = {
     "OK",
     "DATATRF"
 };
-#define maple_resp_name_cnt (sizeof(maple_resp_names)/sizeof(char *))
+#define maple_resp_name_cnt ((int)(sizeof(maple_resp_names)/sizeof(char *)))
 
 /* Return a string representing the maple response code */
 const char * maple_perror(int response) {
